+++ src/com/iesam/huellas/domain/Adoptante.java
@@ -1,35 +1,10 @@
 package com.iesam.huellas.domain;
 
-public class Adoptante {
-    private String id;
-    private String name;
-    private String dni;
+public class Adoptante extends Persona{
+
     private String fechaNacimiento;
     private String direccion;
 
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDni() {
-        return dni;
-    }
-
-    public void setDni(String dni) {
-        this.dni = dni;
-    }
 
     public String getFechaNacimiento() {
         return fechaNacimiento;
