+++ src/com/iesam/huellas/domain/Persona.java
@@ -1,0 +1,40 @@
+package com.iesam.huellas.domain;
+
+public class Persona {
+    protected String id;
+    protected String name;
+    protected String apellidos;
+    protected String dni;
+
+    public String getId() {
+        return id;
+    }
+
+    public void setId(String id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getApellidos() {
+        return apellidos;
+    }
+
+    public void setApellidos(String apellidos) {
+        this.apellidos = apellidos;
+    }
+
+    public String getDni() {
+        return dni;
+    }
+
+    public void setDni(String dni) {
+        this.dni = dni;
+    }
+}
