+++ src/com/iesam/huellas/domain/Socio.java
@@ -1,0 +1,13 @@
+package com.iesam.huellas.domain;
+
+public class Socio extends Persona{
+    private String cuentaBancaria;
+
+    public String getCuentaBancaria() {
+        return cuentaBancaria;
+    }
+
+    public void setCuentaBancaria(String cuentaBancaria) {
+        this.cuentaBancaria = cuentaBancaria;
+    }
+}
